Remove duplicate code, move functions of little use to better places.
