+++ package.json
@@ -7,7 +7,8 @@
   },
   "dependencies": {
     "express": "^4.16.2",
-    "express-sslify": "^1.2.0"
+    "express-sslify": "^1.2.0",
+    "querystring": "^0.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-beta.36",
